[Refactor]: Move per_op_workers parse/validate to shared interfaces
Promote `parse_per_op_workers` from mooncake/pybind.cpp to
connector_pybind_utils.h (C++) and `_validate_per_op_workers` /
`_parse_per_op_workers_from_dict` from MooncakeStoreL2AdapterConfig
to L2AdapterConfigBase (Python), so all backends can reuse them.
