+++ csrc/storage_backends/connector_pybind_utils.h
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 #pragma once
 
+#include "connector_base.h"
 #include "connector_interface.h"
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
@@ -155,6 +156,24 @@ auto bind_drain_completions() {
     return result;
   };
 }
+inline WorkerPoolConfig parse_per_op_workers(const py::object& obj) {
+  if (obj.is_none()) {
+    return {};
+  }
+  py::dict d = obj.cast<py::dict>();
+  std::unordered_map<std::string, int> per_op;
+  for (auto& [key, value] : d) {
+    std::string key_str = key.cast<std::string>();
+    int count = value.cast<int>();
+    if (count <= 0) {
+      throw std::runtime_error("per_op_workers['" + key_str +
+                               "'] must be a positive integer");
+    }
+    per_op[key_str] = count;
+  }
+  return {std::move(per_op)};
+}
+
 }  // namespace pybind_utils
 }  // namespace connector
 }  // namespace lmcache
