+++ csrc/storage_backends/mooncake/pybind.cpp
@@ -10,29 +10,6 @@
 
 namespace py = pybind11;
 
-namespace {
-
-lmcache::connector::WorkerPoolConfig parse_per_op_workers(
-    const py::object& obj) {
-  if (obj.is_none()) {
-    return {};
-  }
-  py::dict d = obj.cast<py::dict>();
-  std::unordered_map<std::string, int> per_op;
-  for (auto& [key, value] : d) {
-    std::string key_str = key.cast<std::string>();
-    int count = value.cast<int>();
-    if (count <= 0) {
-      throw std::runtime_error("per_op_workers['" + key_str +
-                               "'] must be a positive integer");
-    }
-    per_op[key_str] = count;
-  }
-  return {std::move(per_op)};
-}
-
-}  // namespace
-
 PYBIND11_MODULE(lmcache_mooncake, m) {
   py::class_<lmcache::connector::L1RegistrationConfig>(m,
                                                        "L1RegistrationConfig")
@@ -48,7 +25,8 @@ PYBIND11_MODULE(lmcache_mooncake, m) {
                        py::object per_op_workers) {
              return new lmcache::connector::MooncakeConnector(
                  std::move(config), num_workers, l1_registration,
-                 parse_per_op_workers(per_op_workers));
+                 lmcache::connector::pybind_utils::parse_per_op_workers(
+                     per_op_workers));
            }),
            py::arg("config"), py::arg("num_workers"),
            py::arg("l1_registration") =
