+++ lmcache/v1/distributed/l2_adapters/config.py
@@ -256,6 +256,60 @@ class L2AdapterConfigBase(ABC):
         persist_enabled = bool(d.get("persist_enabled", True))
         return PersistConfig(persist_enabled=persist_enabled)
 
+    @staticmethod
+    def _validate_num_workers(raw: object) -> int:
+        """Validate and return a positive integer worker count.
+
+        Raises:
+            ValueError: If ``raw`` is not a positive integer.
+        """
+        if isinstance(raw, bool) or not isinstance(raw, int) or raw <= 0:
+            raise ValueError("num_workers must be a positive integer")
+        return raw
+
+    @staticmethod
+    def _validate_per_op_workers(
+        per_op_workers: dict[str, int] | None,
+    ) -> dict[str, int] | None:
+        """Validate per-operation worker counts (``None`` is a no-op).
+
+        Raises:
+            ValueError: If any worker count is not a positive integer.
+        """
+        if per_op_workers is None:
+            return None
+        for key, value in per_op_workers.items():
+            if isinstance(value, bool) or not isinstance(value, int) or value <= 0:
+                raise ValueError(
+                    f"per_op_workers[{key!r}] must be a positive integer, got {value!r}"
+                )
+        return per_op_workers
+
+    @staticmethod
+    def _parse_per_op_workers_from_dict(
+        d: dict[str, object],
+    ) -> dict[str, int] | None:
+        """Parse ``per_op_workers`` from a raw configuration dict.
+
+        Returns ``None`` if the key is absent.
+
+        Raises:
+            ValueError: If the value is not a dict of integers.
+        """
+        raw = d.get("per_op_workers")
+        if raw is None:
+            return None
+        if not isinstance(raw, dict):
+            raise ValueError("per_op_workers must be a dict")
+        per_op_workers: dict[str, int] = {}
+        for k, v in raw.items():
+            if isinstance(v, bool) or not isinstance(v, int):
+                raise ValueError(
+                    f"per_op_workers[{k!r}] must be an integer, got {type(v).__name__}"
+                )
+            per_op_workers[str(k)] = v
+        return per_op_workers
+
     @classmethod
     @abstractmethod
     def from_dict(cls: type[T], d: dict) -> T:
